+++ packages/core/src/actions/getFeeHistory.ts
@@ -27,7 +27,7 @@ export function getFeeHistory<
   config: config,
   parameters: GetFeeHistoryParameters<config, chainId>,
 ): Promise<GetFeeHistoryReturnType> {
-  const { chainId } = parameters
+  const { chainId, ...rest } = parameters
   const client = config.getClient({ chainId })
-  return getAction(client, viem_getFeeHistory, 'getFeeHistory')(parameters)
+  return getAction(client, viem_getFeeHistory, 'getFeeHistory')(rest)
 }