+++ packages/core/src/exports/query.ts
@@ -91,6 +91,15 @@ export {
 } from '../query/getBlockTransactionCount.js'
 
 export {
+  type GetBytecodeData,
+  type GetBytecodeOptions,
+  type GetBytecodeQueryFnData,
+  type GetBytecodeQueryKey,
+  getBytecodeQueryKey,
+  getBytecodeQueryOptions,
+} from '../query/getBytecode.js'
+
+export {
   type GetConnectorClientData,
   type GetConnectorClientOptions,
   type GetConnectorClientQueryFnData,
@@ -136,6 +145,15 @@ export {
 } from '../query/getEnsResolver.js'
 
 export {
+  type GetEnsTextData,
+  type GetEnsTextOptions,
+  type GetEnsTextQueryFnData,
+  type GetEnsTextQueryKey,
+  getEnsTextQueryKey,
+  getEnsTextQueryOptions,
+} from '../query/getEnsText.js'
+
+export {
   type GetFeeHistoryData,
   type GetFeeHistoryOptions,
   type GetFeeHistoryQueryFnData,
@@ -152,6 +170,24 @@ export {
   getGasPriceQueryKey,
   getGasPriceQueryOptions,
 } from '../query/getGasPrice.js'
+
+export {
+  type GetProofData,
+  type GetProofOptions,
+  type GetProofQueryFnData,
+  type GetProofQueryKey,
+  getProofQueryKey,
+  getProofQueryOptions,
+} from '../query/getProof.js'
+
+export {
+  type GetStorageAtData,
+  type GetStorageAtOptions,
+  type GetStorageAtQueryFnData,
+  type GetStorageAtQueryKey,
+  getStorageAtQueryKey,
+  getStorageAtQueryOptions,
+} from '../query/getStorageAt.js'
 
 export {
   type GetTokenData,
@@ -179,6 +215,15 @@ export {
   getTransactionCountQueryKey,
   getTransactionCountQueryOptions,
 } from '../query/getTransactionCount.js'
+
+export {
+  type GetTransactionReceiptData,
+  type GetTransactionReceiptOptions,
+  type GetTransactionReceiptQueryFnData,
+  type GetTransactionReceiptQueryKey,
+  getTransactionReceiptQueryKey,
+  getTransactionReceiptQueryOptions,
+} from '../query/getTransactionReceipt.js'
 
 export {
   type GetWalletClientData,