+++ packages/react/src/hooks/useCall.test-d.ts
@@ -2,7 +2,7 @@ import { type CallReturnType } from 'viem'
 import { expectTypeOf, test } from 'vitest'
 import { useCall } from './useCall.js'
 
-test('select data', async () => {
+test('select data', () => {
   const result = useCall({
     query: {
       select(data) {