+++ packages/react/src/hooks/useCall.ts
@@ -25,7 +25,12 @@ export type UseCallParameters<
 > = Evaluate<
   CallOptions<config> &
     ConfigParameter<config> &
-    QueryParameter<CallQueryFnData, CallErrorType, selectData, CallQueryKey>
+    QueryParameter<
+      CallQueryFnData,
+      CallErrorType,
+      selectData,
+      CallQueryKey<config>
+    >
 >
 
 export type UseCallReturnType<selectData = CallData> = UseQueryReturnType<