+++ playgrounds/next/next.config.mjs
@@ -1,0 +1,14 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  experimental: {
+    externalDir: true,
+  },
+  webpack(config) {
+    config.resolve.extensionAlias = {
+      '.js': ['.js', '.ts'],
+    }
+    return config
+  },
+}
+
+export default nextConfig