+++ playgrounds/next/src/wagmi.ts
@@ -1,0 +1,28 @@
+import { http, cookieStorage, createConfig, createStorage } from 'wagmi'
+import { mainnet, sepolia } from 'wagmi/chains'
+import { injected, walletConnect } from 'wagmi/connectors'
+
+export const config = createConfig({
+  chains: [mainnet, sepolia],
+  connectors: [
+    injected(),
+    walletConnect({
+      projectId: 'bd4997ce3ede37c95770ba10a3804dad',
+      showQrModal: false,
+    }),
+  ],
+  storage: createStorage({
+    storage: cookieStorage,
+  }),
+  ssr: true,
+  transports: {
+    [mainnet.id]: http(),
+    [sepolia.id]: http(),
+  },
+})
+
+declare module 'wagmi' {
+  interface Register {
+    config: typeof config
+  }
+}